Add getter functions to Kademlia table.
Added a `find` method to locate a node with a particular key in the table. This is useful for implementing protocols, not for querying by remote nodes (see `lookup` for that). Also added the `peers` method to produce a `Seq` of all nodes in the table. This is used produce targets for broadcast messages, for example.

Added a few tests to sanity-check these functions.