Add low-level unicast (UDP) transport.
This is the very lowest level of the communications layer. There will eventually be one of these for each transport, starting with UDP. There is an abstraction for sending and receiving an array of bytes and an abstraction for nodes or peers in a network.

The UDP Comm class does not handle request-response, which I've put at a higher level. There is a dead-simple framing protocol to get right-sized byte arrays from datagrams. There is not quite enough handling of I/O exceptions, yet.