Add kademlia table and PING/PONG.
This PR builds on the `UnicastComm` object to enable overlaying a simple network via UDP. Overlaying network `net` and attaching to a remote node `remote` can be as simple as:
```
import coop.rchain.comm._
val net = UnicastNetwork(NodeIdentifier("local node".getBytes), Endpoint("localhost", 33313, 33313))
val remote = new ProtocolNode(NodeIdentifier("alien node".getBytes), Endpoint("localhost", 33314, 33314), net)
remote.ping
```
which "homes" this view of the network at udp://localhost:33313 and pings the node at udp://localhost:33314.

The `UnicastNetwork` class builds a kademlia routing table using only PING and PONG messages. It extends the `ProtocolHandler` trait with UDP transport versions of the slightly higher-than-base-level comms functions for request-response (`roundTrip`) and `broadcast` (though this PR doesn't use `broadcast`). A `UnicastNetwork` continually receives network traffic in a dedicated thread. The tightness of the receive loop is a constant 500ms currently, controlled by a socket timeout down in the `Comm` object, and it should be parameterized in the future. Messages are handled in the order received.

Since UDP is (for all intents and purposes) connectionless, the `UnicastNetwork` handler class simulates synchronous request-response by blocking in `roundTrip` on a promise of a response. These promises are stored in a `pending` map, keyed by some information in the original message. There is a timeout in `roundTrip`, distinct from the receive timeout above, that can be tailored per call. After the timeout tolls, the promise is failed. The promise succeeds when the receiving thread is able to match an incoming message with one in the pending map.

A node's Kademlia table in this version of the protocol simply includes every remote node it learns of (because that remote node has sent it a PING).

Ultimately, all messages are serialized protocol buffers on the wire. The protocol classes are insulated from this by the `ProtocolMessage` object, which is responsible for parsing arrays of bytes into Scala objects and thereby for some measure of syntactic validation of messages. This object is pretty messy right now, and it will need a good going-over when it does more than PING and PONG; I think something more elegant will fall out of it when there's more to it. If anyone has a suggestion for how to structure it at this point, I'm all ears.

CORE-29 CORE-23 #comment Higher-level Kademlia-lite protocol