Clean up ProtocolMessage and use Either not Try (#27)
* Add kademlia table and PING/PONG.

This PR builds on the `UnicastComm` object to enable overlaying a simple network via UDP. Overlaying network `net` and attaching to a remote node `remote` can be as simple as:
```
import coop.rchain.comm._
val net = UnicastNetwork(NodeIdentifier("local node".getBytes), Endpoint("localhost", 33313, 33313))
val remote = new ProtocolNode(NodeIdentifier("alien node".getBytes), Endpoint("localhost", 33314, 33314), net)
remote.ping
```
which "homes" this view of the network at udp://localhost:33313 and pings the node at udp://localhost:33314.

The `UnicastNetwork` class builds a kademlia routing table using only PING and PONG messages. It extends the `ProtocolHandler` trait with UDP transport versions of the slightly higher-than-base-level comms functions for request-response (`roundTrip`) and `broadcast` (though this PR doesn't use `broadcast`). A `UnicastNetwork` continually receives network traffic in a dedicated thread. The tightness of the receive loop is a constant 500ms currently, controlled by a socket timeout down in the `Comm` object, and it should be parameterized in the future. Messages are handled in the order received.

Since UDP is (for all intents and purposes) connectionless, the `UnicastNetwork` handler class simulates synchronous request-response by blocking in `roundTrip` on a promise of a response. These promises are stored in a `pending` map, keyed by some information in the original message. There is a timeout in `roundTrip`, distinct from the receive timeout above, that can be tailored per call. After the timeout tolls, the promise is failed. The promise succeeds when the receiving thread is able to match an incoming message with one in the pending map.

A node's Kademlia table in this version of the protocol simply includes every remote node it learns of (because that remote node has sent it a PING).

Ultimately, all messages are serialized protocol buffers on the wire. The protocol classes are insulated from this by the `ProtocolMessage` object, which is responsible for parsing arrays of bytes into Scala objects and thereby for some measure of syntactic validation of messages. This object is pretty messy right now, and it will need a good going-over when it does more than PING and PONG; I think something more elegant will fall out of it when there's more to it. If anyone has a suggestion for how to structure it at this point, I'm all ears.

CORE-29 CORE-23 #comment Higher-level Kademlia-lite protocol

* Remove spurious return.

* Update Try[Boolean] to Try[Unit]

Feedback from @tymm

* Too lazy!

Feedback on #22 from @tymm.

* Add some clarity in for comprehension.

Also, scalafmt did some violence.

* Add lookup and lookup response.

This completes as much of the Kademlia protocol as we need in order to create and maintain a network. Particulars:

* Moved return header of protocol buffers up to top-level
* Adjusted encode/decode return types as suggested by @tymm in #22
* scalamft'd the kademlia test cases

* Update for PR feedback.

* Add return types
* Simplify `toPeer`

* Explicitly return Unit instead of ignoring value.

* Moved many functions from ProtocolMessage object to trait.

This keeps low-level protocol buffers from leaking to higher level protocol handler code.

Also removed some unnecessary `Option` wrapping.

* Removed some unneeded functions.

* Adjusted line breaks for consistency.

* Added missing return types.

* Add a couple comments.

* Updates for PR commentary.

* Remove `???` case that can't be reached.
* Remove catch-all-exceptions in favor of `Try` block.

* Changed Try to Either.

Mostly.

* Remove commented-out code.

* Really resolve conflicts. Sheesh.

* Remove commented-out import statement.

* Update for PR comments.

Change `Throwable` to `Exception` and fix stale comments.

* Collected all errors under CommError

Not sure this is the best breakdown of these errors, since calling code isn't yet written, but this structure does have the virtue of letting folks see all the errors in one place.

Also simplified some things per PR comments.