Build each subproject, not just rholang, rosette
In the Nov 3 "CI configuring notes" design, only directories changed
in the last commit are built. But if a PR has a number of commits, the
last of which is only a change to the top-level README, any
subdirectories wouldn't be re-built. Some searching suggests this is
an unsolved problem:

  - https://stackoverflow.com/questions/39596177/monorepo-travis-and-matrix-builds-how-do-i-build-a-specific-path-only-when-it
  - https://github.com/travis-ci/travis-ci/issues/6301

So let's just build everything.

Also:
  - use `test.sh` in preference to `build.sbt` if present
    to support
    - the C++ VM
    - the BNFC step etc. in rholang
  - use `-Dsbt.log.noformat=true` to avoid clutter in logs