Complete the grammar and resolve parse amibiguity.
The parse ambiguity stemmed from using @ for both lifting a process type to a
namespace and using it to match a specific name.

Interestingly, this means @Nil and @@Nil are valid namespaces, both describing
the same set {@Nil} of names. @@Nil lifts the Nil pattern to a namespace, and
@Nil lifts the Nil process to a namespace. @@true and @true are not the same
namespace, however. @@true is all names, and @true is the namespace that matches
the quoted process true.

If we feel that the @@ is too visually similar to @, we can pick a different
symbol.