Add README and bump version.
The README still needs the bootstrap node address, which we're working on today. I'd like to keep this part simple, but let me know if it needs more detail (like _how_ to check out the repo, what directory to change into to build, and so on).

Also, removed a `println` and added a Java option to the Dockerfile's runtime environment to try to stop Java from picking an IPv6 address to use.

Bumped library version number.