Rholang: Use a simpler visitor.
This is a rather invasive change, but it will make the compiler much easier to
understand. The compiler was written using a FoldVisitor, which by default
combines results from the leaves upward, along with an argument.

For all of the visit methods that override the default FoldVisitor behavior, the
combine method was always called with the arg parameter from the visit method.
In turn, the arg parameter was always Here(). Reading through Here(), and the
combine method, this means that combine was just returning the tree being
combined with arg.

For the visit methods that were not overridden, I don't believe they were
correct. For example:

Given the test program:
{
  for (_ <- @Nil) { print("Hello") } | @Nil!(Nil)
}

It compiled to:
(block
  (let [[
      [[Rholangf5f069da6bde] [Rholange787f4c3d260]]
      (consume t [#niv] [**wildcard**] ['Right(ContextVarRholanga86a9a57a19a)] #f)]]
      ( (proc [[Right(ContextVarRholanga86a9a57a19a)]]
          (seq (print "Hello") (display #\\n))) [Rholange787f4c3d260]))
  (produce t #niv **wildcard** #niv))

Which is odd in some obvious ways: 'Right' should not be part of the compiler's
output. Given that the token contract relies on the wildcard, I've added that
here. It's small.

This occurs because of how the combine method is written.

This is the first commit in a series of drastic cleanups. If no
context-combining actually occurs, then we can get rid of the zippers and the
location contexts, and instead just use namespaced trees. Given that Here() was
the only use of Context variables, we can remove the Context/Lang distinction.

Right now arg is empty, but in the future it makes sense to use it as an
environment parameter. We really should fail to compile if an unused variable is
mentioned, but right now we don't.