Rholang: Remove most of the zipper code.
The zipper code was unused, and in the interest of making the compiler more
comprehensible, I'm trying to remove as much unused code as possible. Remove all
references to the context part of the zipper library.

The visitors for InputBind appeared to need it, but the contexts that contain
InputBind's don't call visit directly, so that code was removed and replaced
with an internal compiler error.