Make bootstrapping the default.
The usual use of this software will be to attach to the main rchain network, so that's now the default. A stand-alone node, which will not attempt to bootstrap from anywhere, may be brought-up with `--standalone` or `-s`. It is still allowable to specify the node from which to bootstrap, so bringing up a "private" network is still as easy.