Add uPnP port mapping. (#101)
* Add uPnP library.

* Update version number.

Also add rchain/ repo name to docker image name.

* add upnp stuff

* Removed stale message type.

Adding simplified uPnP.

* Update docker image name.

* Comment.

* Make bootstrapping the default.

The usual use of this software will be to attach to the main rchain network, so that's now the default. A stand-alone node, which will not attempt to bootstrap from anywhere, may be brought-up with `--standalone` or `-s`. It is still allowable to specify the node from which to bootstrap, so bringing up a "private" network is still as easy.

* Updated README.md.

Also removed/changed some print statements.

* Updating for uPnP and cleaning up language.

It's still confusing.