Add a few simple, colorful tests. (#130)
* Add a few simple, colorful tests.

This file is just a starting point for writing tests against the C++ rosette implementation. This isn't meant to be an exhaustive battery; rather, it let me (1) play around with rosette macros and (2) understand the rosette environment a little.

I've also made executable the test-running script and added the tests there. Not sure that's the right thing to do.

* Remove silly debugging prints.