Hacks to try to add regression tests for segfaults.
For example, before something ROS-304 is fixed in #137, we would get this:

```
Summary:
Regression failure (exit code 139) on rbl/rosette/tests/segfaults/regression-ros-304.rbl
```

and after:
```
Summary:
All good.
```

This is super hacky, since I don't see any way to recover from either the segfault or even a reader or parse error in Rosette. Maybe this is where "monitors" come in, but that part of the documentation isn't.