Add simple logger.
This is a simple logging facility. There is a global log level (`*log-level*`) controlled initially by the `flag-verbose` parameter. There are four predefined log level "constants": `'ERROR`, `'WARNING`, `'INFO`, and `'DEBUG`, in order from most to least severe.

Expressions of the form
```
  (log:display 'ERROR "This is a " 'fancy " log message.")
```
emit a newline-terminated, timestamped log line to the standard error stream like:
```
  2018-01-16 18:35:35 ERROR: This is a fancy log message.
```
if the given level is at least as severe as the global log level.

The global log level may be manipulated, though it seems rude:

```scheme
  rosette> (define *log-level* 'ERROR)
  '*log-level*
  rosette> (log:display 'WARNING "This is a warning!")
  rosette> (define *log-level* 'WARNING)
  '*log-level*
  rosette> (log:display 'WARNING "This is a warning!")
  2018-01-16 18:41:17 WARNING: This is a warning!
```

Also included is a primitive `log-time-string` which formats local time into a string and returns it.