Rholang: Add Grammar for Rholang Mercury spec.
The grammar is missing a few things, but is fairly complete.
Verified that the lexer and parser are happy with the files produced by bnfc.
The grammar ends up shorter because we use the same production for processes and
patterns. A free variable in a process used as a pattern is a binder.