Rholang: Add initial work on normalizer.
Normalizer will work in 2 parts:
Part 1 will "compile" the source down to a partially normalized form, and part 2
will sort the subcomponents of all subproccesses according to an order.

This is the beginning of part 1: it adds some visitor classes to start producing
a normalized form from the syntax, and tests! to make sure that that the few
productions for which we have visitors are working.

Adds scalatest as a testing dependency.