Refactor Network constructor to take PeerNode (rather an arbitrary String)
Parsing String should not be concenr of `Network` constructor. A
dedicated `PeerNode` should be created by the client and any error
handling of peerNode creation should happen out side of this scope.

Also this change did not require any tests to be modified which is
troubling to say the least.

Further work should consider proper error handling in main method as
well as removing `connect` method that take `String` as an argument.