Add WartRemover with initial Wart.FinalCaseClasses
This is initial configuration of WartRemover. I've noticed that many
classes should not be case classes or if they are, they should be
marked as final. Added WartRemover to do the work for me.

This configuration will be extended in the future, while we keep on
improving towards more FP code base