Rholang: Normalize wildcard to a free variable.
We want the following to match:
for (x <- y) { P } // x is not used in P
for (_ <- y) { P }

In order to do this, we treat wildcards as if they were a bound variable. We
don't insert them into the binding environment used for tracking, but we bump
the counters used for DeBruijn levels. There's a subtle point coming when I get
to pattern matching, which is that the free variables in a pattern act as
binders, now including wildcards. It should work to walk the free map and bump
the levels there by the 'next' amount in the target map. Then bump the next amount
of the target map by the 'next' amount of the free map.