Rholang: Normalizer: Collapse eval/quote and quote/eval pairs.
In Rho Calculus, you have to pay attention to bound/unboundness for quote eval
substition. Here we use variables for variables, and so any names that appear
are not bound. (They can't be by construction). Because of this, quote and eval
are exact inverses. When normalizing, we remove quote/eval and eval/quote pairs.
We don't need to recurse because we remove at both sides. There's a simple
structural proof that we won't end up with any quote/eval or eval quote pairs if
we remove them whenever we see them.