adjust compiler settings, enabling deprecation warnings
It is helpful to know when we are using deprecated features in our
code, so we should enable this warning.  Since we are not using
"-Xfatal-warnings" (yet), it will not prevent us from building code with
deprecated features, it will merely warn us, which is good.

This commit also re-adds "-Xfatal-warnings" to the set of options that
will be disabled when using the REPL (aka "console").  It effectively
changes nothing right now, but it defends against a future situation
where we enable "-Xfatal-warnings" and forget to change this setting,
resulting in annyoing behavior at the REPL.  Plus, commented-out things
are smelly.