Rholang: ADT: Add locally free bitset.
This will allow us to do matches of terms with environments. It turns out that a
free variable is not the same as true, because you can only bind things that
don't have any locally free variables. For a fully substituted term, all we need
is the lowest value used in a term to figure out if we can match, but with an
environment, we need to know what the lowest would be once all the variables in
the environment are substituted. Hence, a bitset.

This will likely complicate using a protobuf. On the plus side, It's info that
could be re-calculated on deserialize, or we can serialize the bitset as bytes
in the proto.