Remove models module and prepare CBOR serialization for rholangADT
The models module contains protobufs that generated case classes that
were going to replace the current rholangADT. Unfortunately, protobufs does not support
bitset fields, which we now use in the rholangADT to keep track of the
minimum level boundvar - see https://github.com/rchain/rchain/pull/354.

Our options are as follows:

1) Keep both the rholangADT and the protobufs even though they are
almost exactly the same.
2) Serialize the bitsets to binary and deserialize back when doing a
bitset operation. This implies that for each OR operation done on a
bitset, we have to convert to a bitset, do the OR operation, and then
convert the result back to binary.
3) Throw away protobufs and add custom serializer wrappers around the rholangADT.

We are going with option 3 - the last one.