Models: Update RhoTypes.proto with updates to rholangADT
This includes adding bitsets through the use of a custom type and
a TypeMapper from ByteString <=> BitSet. The tests use
ScalacheckShapeless which automatically generate arbitrary ADTs based on
the generated protobufs case classes. This is a really powerful testing
feature.

This supersedes #360 and is part of
https://rchain.atlassian.net/browse/RHOL-142.

Thanks to @henrytill for his help in coding these changes.