Rholang: Basic Matcher functionality.
Add a basic implementation of the matching functionality.
Things that work:
Expressions in parallel, including lists for argument matching.
Recursively descends into lists/tuples.

Things that need to be added:
Matching for anything besides sends and expressions.

This is making me realize that we need to use DeBruijn Indices. That will make
capture far more simple.