Rholang: Substitute: Use Debruijn Indices.
Debruijn indices for substitution don't require re-naming. It does require
actually creating an environment class. We need to be able to shift the keys of
the map in 2 different cases. The first is when we add something to the
environment. The second is as we substitute and move inside of a binding
construct.