Rholang: Reduce: add reduction for match and tests.
Matching involves evaluating any expressions in the target, and then for now,
doing full substitution of the target. Eventually we'll be able to hand the
matcher an environment so that we don't have to do substitution before matching.

As a part of getting this to work correctly, I made send fully substitute the
data. This is also something that we would eventually like to remove.