Integrate REPL into Node
There is a lot of code duplication between RholangCLI and this file,
so we might want to clean that up. The Node REPL (this one) uses
the LMDB store and the RholangCLI REPL uses the InMemoryStore.

I've noticed the node just crashes when it takes more than 5 seconds
instead of looping. There is probably a lot of fine tuning to be done.