Rholang: Normalizer: Add test for freeCount not including wildcards.
Previously we were counting wildcards as free variables. That changed when we
realizid that wildcard matching was more powerful than free variable capture. We
didn't have any tests to make sure that we weren't rolling up wildcards into the
free variable count. Doing this makes the counts wrong for the evaluation and
leads to incorrect results.