Rholang: Reducer: Allow for some expressions to evaluate to Pars.
Evaluate those expressions to pars at the correct time.
Most expressions only evaluate to other expressions, but variables and method
calls may produce actual terms. This means that we have to handle them in a
slightly different way when we are evaluating expressions for a term to be used.
For example, we do this subtly different evaluation for data being sent on a
channel. This means that you can select the nth process from a list on send, and
that process does not have to be just an expression.