Rholang: Remove freeCount from the places where it isn't necessary.
We were using freeCount for 2 things: in most cases, we were just checking that
it was non zero. In a few cases, we needed the actual number. FreeCount has been
added to those places where we needed the actual number, and instead we roll the
non-zeroness into the connectiveUsed boolean.