Rholang: Matcher: Make the matcher parameterized in 2 types.
This is necessary for logical connetives where what we're matching is a Par, and
what we're matching against is a logical connective.

Another preparatory change is adding the option of writing a spatial matcher in
a non-deterministic fashion. The matcher for par is now written that way, and
I'll be using that to allow backtracking in how connectives are satisfied.