Rholang: Matcher: Add logical connectives.
This allows the matching of And, Or, and Not for complicated spatial types. Note
that using Not is expensive at the moment, so we may want to add some things
that are expressible with Not as expressions in themselves. Two examples are
~Nil and ~{~Nil | ~Nil} /\ ~Nil. The first being any non-nil term and the second
being any atomic term that cannot be split into 2 things in parallel. The
matching uses backtracking, but we try to be smart and compute bounds on how
many of each atomic term must be in the piece that we match, and how many must
be left for the remaining logical connectives.