Create rholang.k
K framework for rholang. The next major hurdle to overcome is pattern-matching in the listening "for" construct. Right now, we can only listen for names in 1-tuples. We should be able to listen for an n-tuple of any combination of patterns.

Pattern-matching will likely use a combination of type theory and recursive functions; my goal is to make it elegant and simple, although this may not be possible.

There's also no match construct, which is the hurdle to be overcome after the pattern-matching is done.