Rholang: Make eval an expression.
This fixes a bug we had when trying to evaluate eval's inside of expressions. It
should be the case that
for(@x <- y) {z!(x*x)}
behaves the same as:
for(x <- y) {z!(*x * *x)}
but the latter was not working correctly. The best solution to that is to just
make eval another expression, and basically make it do the same as an EVar. Most
of the delta here is just necessary changes to make that happen.