Update rholang.k
This is the first major update of the rholang k-framework. Most significantly, this framework now supports pattern-matching in the listening constructs (i.e. for( pattern <- channel){ process }). The pattern-matching is not perfect; there are a few defects which are documented in the file. Most notably, one can match with a process that has a locally free variable in it, resulting in a free variable in the executed process. Once this, and a couple other things, are taken care of, the framework will fully support pattern-matching.

By far the most involved part of this was matching patterns in which the order of the parallel processes written does not matter. We have to systematically go through possible matches. This is (hopefully) well-explained in the documentation of the document.

I am aware that there are probably many redundancies and inefficiencies; please do let me know if you find any, or if you think of ways to simplify this process. I'm hoping that, as I move forward, new insights will come.

This document is split up into sections, the biggest of which correspond to pattern matching and pattern matching parallel processes.

A new update will soon come with added functionality.