CORE-592: rspace: restructure to support multiple impls.
To run RSpace in "replay" mode, we will need to support multiple
implementations with different semantics for produce and consume.

This commit moves the core functions into a class called RSpace which
implements the ISpace interface.  An instance of PureSpace, which
contains the "pure" API, can be constructed from anything that
implements the ISpace interface.

In the context of "replay" mode, we will construct the Reducer with a
PureSpace that wraps an (as of yet unwritten) ReplayRSpace.