Rholang: Handle interpreter errors by catching and logging them.
Interpreter errors occur in parallel with other processes. Because the ordering
is paralell and non-deterministic, in the absence of some transaction model, all
the succeeding processes could have already run when a process with an error
happens to run. Because this is a valid semantic ordering, this is the order
that we choose. We collect the errors so we can report to the user when
something goes wrong.