Introduce LogSource to extend functionality of logger
Uptil now logger was logging with source set to "main". This was
problematic when working with logs as each message
loged (INFO/WARN/ERROR/DEBUG) were listed as if they were coming from
the same file.

With this change this is no longer a case. Each class that needs to
use Log[F] has to provide an evidence (only once) for a
`LogSource`. This way each time Log[F] method is called, the logger
can be created with the source set correctly