Add run subcommand in systemd service description
When trying to run rnode as systemd process, system was not booting
up. This commit is fixing that. There is a bigger question however,
how users that use rnode can modify avaliable flags when running rnode
as systemd process. For now they can only start rnode with fixed set
of flags. This needs to be addressed.