Add CommError to CommunicationResponse.NotHandled
When message was not handled by the receiver, it most likely is for
some reason. The caller will get an information why the message was
not handled.

Currentyl TLResponse.InternalServerError holds a String, it should
eventually hold CommError (if possible), if not some equivalent form
with CommError can be build from