Rholang: Use Blake2b512Random for deterministic unforgeable names.
Because we want to avoid ordering issues, we can't update the state of a
replicated receive so that it has a different random state, and we definitely
can't have replicated receives re-use random state and generate the same names
each time they are awoken. Instead, we store random state with sends and then
merge the random state of the send and the receive after a comm event.

Note that this is currently broken for infinite loops where a replicated send
meets a replicated receive. We might solve that by having replicated sends and
receives annihilate each other, and return an infinite-loop flag from the
tuplespace. By pushing them both out, we can manipulate the state of the ongoing
infinite loop.

For casper, we hash the DeployString, which must always be unique, in order to
seed the Random number generator. If any of the fields were pre-hashed, it would
be more efficient to use that, but at the moment, they aren't.