Rholang: Make locallyFree reflect the top level.
locallyFree was always redundant in patterns because they were never going to be
executed or be matched. Now it is the top-level locallyFree cache. This is
necessary because of patterns. Because we use equality comparisons in matches
without connectives, we need to ignore locallyFree when doing equality
comparisons. Hence locallyFree. Similarly in ParMap and ParSet, we also ignore
locallyFree.
We added a depth counter to HasLocallyFree.locallyFree.
HasLocallyFree.locallyFree relies on the cached information, and therefore will
only return the correct result if asked about the top-level. This is fine,
because the top-level is the only place that it matters.