Make MatchTest capture comparison errors more clear by printing rholang
Most term comparison failures are due to substantial differences in the
captured terms. In such cases, comparing (and showing a diff of) the
corresponding rholang allows for easier and faster understanding of the
failure.

In all cases, the bare terms are compared too, so differences
not reflected in the source code will also be spotted and reported.