Adding new prometheus reporter
The new prometheous reporter does not use internal http server but
feeds that data to the http4s sever.

You might notice that in node.scala we actually fire up TWO https4s
servers (running on 40402[http-port] and 40403[metrics-port]).
Eventually we want to have only one server (running on the http-port),
but because removing the metrics-port might be a breaking change for
the tooling that we have (integration tests, harness tests etc) -
removing that metrics-port will be brought in seperate PR