Handle 'lots of targets, no patterns, remainder' edge case separately
This is a very common case for usage in rspace, and - of course -
a worst case scenario for Maximum Bipartite Match. Having no patterns
and a reminder means running the MBM with #targets RemainderPattern-s,
which means all patterns in MBM match everything, causing each of the
considered patterns 'push back' all the so-far-matched patterns to
further targets. This results in (N over 2) iterations of the algorithm,
(roughly O(N^2)) for an edge case that can be handled in linear time.