Rholang: Remove Channel and Quote from the proto
They aren't necessary, and they complicate working with the protos.

Fixed everywhere they are used, and all tests pass after the removal. After this
commit, you will need to run sbt rholangProtoBuild/clean casper/clean at a
minimum. This a breaking change for the term encoding, which is why I'm trying
to get it in before the release.