Remove depth from listen for name endpoints in node client
Depth doesn't make sense on these endpoints because we are listening
and the responses are diffed by size. Thus if we used default depth 1,
and we got back responses of the same length even though the contents
have changed, the listen endpoint on the client would return empty
even though we would expect it to return the newer contents.