Change streamBlob(Blob) to stream(Packet)
This change is made after conversations that I had with Micheal
Birch. Comm needs to stay totally NOT aware of abstraction that is
using it. We Achieved this by introducing `Packet` and
`PacketHandler`. `Packet` contains arbitrary data and `PacketHandler`
is a typeclass (an interface) for the user of Comm to implement -
giving comm information how to handle incomminb messages that are not
part of the initial RChain Protocol.

Introducing `Blob` was breaking that design concept. It was pushing a
cycle into dependency graph (comm needs to know casper, casper needs
to know comm), which in a long run is highly potential source of
errors and confusion.

Having dedicated method `def stream(peers: Seq[PeerNode], packet:
Packet)` allows us also take adventage of existing logic hidden in
`PacketHandler`. It's going to be Casper to decide if he wants to send
packet normally via `roundTrip`, `send` or `broadcast` (wrapped in
`Protocol`) or if he wants to stream it via `stream`.

Please not that `def stream(Seq[PeerNode], Packet)` is still not
sending messages in stream manner - that is a change to come