Fix non-deterministic test failure caused by lack of reAttempt
There are two problems this PR fixes:

1) We previously removed block hashes that were dependency free that
weren't yet in the block buffer. Thus when the block B actually came in
any block that depended on B might have not been triggered to be added.

2) We forgot that when we added the "processing" BlockStatus that now this
makes it such that some blocks may fail on reAttempt even though they
are dependencyFree. So while it is a bit ugly, we whitelist instead of
blacklist which blocks get successfully added during a reAttempt call.