Completely redo compression in stream
This commit provides new way of dealing with compression. Lz4 is still
being used but we use one that is using C implementation under the
hood. The compression outputs  are interchangalble, you can compress
in Java/Scala and decompress in C - which is a value of its own. It is
potentially faster (copies bytes in memory without asking)