Introduce a way to fetch bootstrap that was used to connect to p2p
If we start in standalone, the bootstrap in conf will return None. If
we started booting up from some --bootstrap, then that PeerNode will
be returned within Option.

*Motivation:*
This change is required by Casper team to provide a better mechanism
to request an approved block. Casper would like to ask bootstrap node
for that information