Introduce queue on the client side of the stream API
After this change, whenever you call stream with given Blob, that
message is pushed on a queue and the code finishes the
execution. There is a seprate consumer of this queue that will then
pop elements from it and send it (in stream fashion) to the
destination.

This way calling `stream` API executes immediatly (in matter of
seconds) even for big number of peers involved.

Next PRs should bring to it:
- back pressure (as much as it is possible)
- retries