boot-p2p: clarify control flow, esp. re bonds (OPS-331)
I wasn't quite sure what the code was doing, so I refactored it to
clarify flow of control and access.

Now that I see what's going on, I think the previous version was
correct: it did add to the bonds file one line for each bonded
validator (according to --sigs).

But perhaps others would agree this is easier to read.

In detail:

  - factor out access to start containers as class DockerImage
  - use classes P2PNetwork, RNode, BootstrapNode, PeerNode
    to clarify control flow and reduce scope of variables
  - refactor generate_validator_private_key as
    KeyPair.encode(create_keypair())
  - factor out argparse boilerplate with docopt
    - add Flags class to handle int args
  - pass authority around explitly, following ocap discipline
    - use pathlib rather than os.path
  - use logging rather than print()
    - add --debug flag
  - prune unused RNODE_CMD