Fix toContent (in stream api) to handle correctly data that was not compressed
The reason for bug was following: When data was received and
colledcted by the server, it then was trying to check if the received
data was compressed before sending. Instead of checking the boolean
flag `compressed`, the server checked if the `contentLength` field was
set. That field however is always set (regardless if the data is
compressed or note), thus it was ALWAYS trying to decompres received
data.

This commit fixes the problem. It correctly checks if flat `compressed
` is set to true and only then attempts the decompression process.

TODO: this bug was not caught by tests. Reason: no test for stream API
when sending small messages (that will not be compressed). This needs
to be addressed. Added CORE-1449 to address that