Modify stream on the server side so that it can deal with arbitrary blocks
Previously incoming blocks were aggregated into a `PartialBlob` object
and then transformed to `Blob` which was handled by Casper. This could
lead to number of problems, mainly OOME. With this new approach, we
collect incomming stream of data by saving chunk by chunk into
file. Once the file is stored, we push just a path on the queue. This
way consumer of the queue handles each blob one at a time, retrieveing
it from a file. This gives as a way to provide fine-grade
mesassurments to deal with blobs of arbitrary size.