Introduce a circut breaker mechanism when handling a stream
When handling a stream (receiving incoming data), handler might decide
to break a circuit (base on how much data was read so far). This way
if node receives a block that is simply to big, it will ignore
it (will not handle it).

Even though the mechanism is in place, currently circuit breaker
accepts any incomming stream (it never breaks). What should be
conditions of it breaking the circuit are a matter of discussion
planned in RCHAIN-2792