Restore assembly plugin 'deduplicate' mergeStrategy
This must've been disabled by setting
`Global / conflictManager := ConflictManager.strict`
a while ago.

With this re-enabled, we pay some more build time, but at the same time
are again sure that if a file in the assembly would be overwritten
by some other file provided by a dependency (or vice-versa), this is
only possible if their contents are exactly equal. So we can ignore it.

It's actually a bit more complex than that.
See https://github.com/sbt/sbt-assembly#merge-strategy,
`defaultMergeStrategy`, and the merge strategy provided by this commit.