- Add ParserError as a subtype of InterpreterError
- Use lexer and parser constructors as entry-point for monadic context.

Since constructor errors are adapted in the lexer and parser smart constructors buildAST only needs to adapt errors thrown during parser.pProc(), hence the adaptError method being moved to the inside of the for-comprehension.

- Remove try/catch block from buildNormalizedTerm.

Since buildAST adapts errors during lexing and parsing, and buildPar only uses normalizeTerm and sortMatch (both of which are abstract over F) it's reasonable to assume that compiler errors to this point have already been sequenced.