Modify MultiParentCasperRef to return default closed over effect
Whole approach of `MultiParentCasperRef` is **at least**
bizzare... Transitions are being made in `PacketHandler`, which sets a
value of a `MultiParentCasperRef`. This does not look as a right
design and should be addressed. In particular, node should be
constantly aware in which state it's in and react accordingly.

For now, in this change, we are giving a visibility **which* method
was used, to trigger `casper not available` output.

We must rememgber that `casper not available` is **not** an
error. This is normal that node might be in transition while receving
different message and command from api. `casper not available` is just
kind remainder to the caller "i'm not ready, yet".